Storing Students Ratings in Database
Adjusted the database in order to be able to store student ratings. Handled errors that would occur if the student had no one to rate. Created an evaluatorID and evaluateeID which would be stored each time a students rates another. (also created a teamID to further label but might be subject to review). Slightly adjusted two frontend files to work with the backend.